ai-engine/requirements.txt
# Dakdan AI Engine - Lead Scoring API
# Install with: pip install -r requirements.txt

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
supabase>=2.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
